+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## Unreleased
+### Added
 - Adds `histogram_to_string` to output histogram plot as string. Thanks to
   @stodoran for suggestion and PR!
+- The `color` option now also supports a single named color, though outputting
+  a warning.
 
 ## [0.21.3] - 2025-07-27
 ### Added